Update our registry-reading patch for modern SCons.
When building with Wine, we monkey-patch the registry-reading
code in SCons to return the path to our installation of MSVC
(since it can't read the registry directly as we're cross
compiling on a Linux host). The function used to take just
a single value; in newer SCons it takes two values (the second
is the registry root). Modify our patch to work when called
either way.